+++ R/enrichment.R
@@ -79,7 +79,8 @@ clusterORA <- function(g,
                        alpha = 1.0,
                        col = COLLAPSE) {
     anL <- getAnnotationVertexList(g, name)
-    cl <- make_clusters(g, as.numeric(vertex_attr(g, alg)))
+    cl <- make_clusters(g, membership=as.numeric(factor(vertex_attr(g, alg))),
+                        algorithm = alg)
     vcnt <- vcount(g)
     forafun <- function(.i) {
         gids<-which(membership(cl) == .i)