+++ R/memrob.R
@@ -45,8 +45,8 @@ getRobustness <- function(gg, alg, conmat) {
         )
     }
     rm <-
-        data.frame(cm = as.numeric(igraph::vertex_attr(gg, alg,
-                                                                V(gg))))
+        data.frame(cm = as.numeric(factor(igraph::vertex_attr(gg, alg,
+                                                                V(gg)))))
     cm           <- data.frame(conmat)
 
     names(cm)    <- rownames(rm)